Small but important change to restore doc name in output. Hasn't fixed everything yet but it's better.

+++ lucene43/test/pitt/search/semanticvectors/integrationtests/RegressionTests.java
@@ -84,16 +84,23 @@ public class RegressionTests {
 
     List<SearchResult> results = Search.RunSearch(FlagConfig.getFlagConfig(searchArgs));
     int rank = 1;
+    boolean found = false;
     if (results.isEmpty()) {
       throw new RuntimeException("Results were empty!");
     } else {
       for (SearchResult result : results) {
         String term = (String) result.getObjectVector().getObject();
-        if (term.equals(targetResult)) break;
+        if (term.equals(targetResult)) {
+          found = true;
+          break;
+        }
         ++rank;
       }
     }
-
+    if (!found) {
+      fail("Failed to get matching results with search terms: " + searchCmd);
+    }
+    
     for (String fn: filesToBuild) assertTrue((new File(fn)).delete());
     return rank;
   }
@@ -102,6 +109,9 @@ public class RegressionTests {
   public void testBuildAndSearchBasicRealIndex() {
     assertEquals(2, buildSearchGetRank("-dimension 200 -luceneindexpath positional_index",
         "-queryvectorfile termvectors.bin -searchvectorfile termvectors.bin peter", "simon"));
+    assertEquals(1, buildSearchGetRank("-dimension 200 -luceneindexpath positional_index",
+        "-queryvectorfile termvectors.bin -searchvectorfile docvectors.bin peter",
+        "test/testdata/John/Chapter_21"));
   }
 
   @Test
@@ -156,6 +166,12 @@ public class RegressionTests {
         new String[] {"termtermvectors.bin", "docvectors.bin"},
         "peter");
     assertTrue(peterRank < 5);
+    int chapter21Rank = positionalBuildSearchGetRank(
+        "-dimension 200 -vectortype real -seedlength 10 -luceneindexpath positional_index",
+        "-queryvectorfile termtermvectors.bin -searchvectorfile docvectors.bin simon",
+        new String[] {"termtermvectors.bin", "docvectors.bin"},
+        "peter");
+    assertTrue(chapter21Rank < 5);
   }
   
   @Test
